refactor(redux): organized the project to follow redux-toolkit file structure pattern
