+++ .prettierrc
@@ -5,5 +5,6 @@
 
   "semi": false,
 
-  "trailingComma": "all"
+  "trailingComma": "all",
+  "printWidth": 80
 }
