+++ src/common/__helpers__/image.tsx
